table: show each expense's "Složka" as well
This is needed for, e.g., "Správa 8/25". That name is used for different
expenses, such as building cleaning, or preparation of garbage trays.
These expenses are sometimes bundled into the same view, and as a
result, the table would show multiple lines with the same name. Let's
make sure that this is correctly separated.