clarify that we're showing partial data
Mr. Haltmar pointed out that the current display doesn't show a list of
expenses over 10 months -- there are some "down payments" for utilities
planned for the rest of the year as well, but on the other hand some
known expenses (like administration fees, etc) for October are not there
yet. Let's remove some potentially misleading statistics and focus on
what we know, and offer some ideas about the last import date, and
provide some instructions for usage.